build : cart

+++ api/src/frameworks/category/category.controller.ts
@@ -1,7 +1,10 @@
 import { Body, Controller, Get, Param, Patch, Post, Delete, UseInterceptors } from "@nestjs/common";
 import { CategoryService } from "./category.service";
-import { ResponseMessage, TransformationInterceptor } from "src/middlewares";
+import { ObjectIdNotFoundException, ObjectIdValidException, ResponseMessage, TransformationInterceptor } from "src/middlewares";
 import { Category } from "./model/category.model";
+import { Types } from "mongoose";
+import { CategoryDTO } from "./dto/categoryDTO.dto";
+import { ValueMessage } from "src/enum/value.enum";
 
 @UseInterceptors(TransformationInterceptor)
 @Controller('api/type-product')
@@ -10,8 +13,7 @@ export class CategoryController {
 
     @Post()
     @ResponseMessage('TypeProduct Create Succesfully')
-    create(@Body() category: Category): Promise<Category> {
-        // console.log(category);
+    create(@Body() category: CategoryDTO): Promise<Category> {
         return this._categoryService.create(category);
     }
 
@@ -23,19 +25,46 @@ export class CategoryController {
 
     @Get(':id')
     @ResponseMessage('TypeProduct fetched Succesfully')
-    findOne(@Param('id') type_id: string): Promise<Category> {
-        return this._categoryService.findOne(type_id);
+    async findOne(@Param('id') type_id: string): Promise<Category> {
+        if (Types.ObjectId.isValid(type_id)) {
+            let category = await this._categoryService.findOne(type_id);
+            if (category) {
+                return category;
+            }
+            else {
+                throw new ObjectIdNotFoundException(type_id, ValueMessage.CATEGORY);
+            }
+        }
+        throw new ObjectIdValidException(type_id, ValueMessage.CATEGORY);
     }
 
     @Patch(':id')
-    // @ResponseMessage('Update TypeProduct Succesfully')
-    update(@Param('id') type_id: string, @Body() category: Category): Promise<any> {
-        return this._categoryService.update(type_id, category);
+    @ResponseMessage('Update TypeProduct Succesfully')
+    async update(@Param('id') type_id: string, @Body() categoryModel: Category): Promise<any> {
+        if (Types.ObjectId.isValid(type_id)) {
+            let category = await this._categoryService.findOne(type_id);
+            if (category) {
+                return this._categoryService.update(type_id, categoryModel);
+            }
+            else {
+                throw new ObjectIdNotFoundException(type_id, ValueMessage.CATEGORY);
+            }
+        }
+        throw new ObjectIdValidException(type_id, ValueMessage.CATEGORY);
     }
 
     @Delete(':id')
-    // @ResponseMessage('Delete User Succesfully')
-    remove(@Param('id') type_id: string): Promise<any> {
-        return this._categoryService.remove(type_id);
+    @ResponseMessage('Delete TypeProduct Succesfully')
+    async remove(@Param('id') type_id: string): Promise<any> {
+        if (Types.ObjectId.isValid(type_id)) {
+            let category = await this._categoryService.findOne(type_id);
+            if (category) {
+                return this._categoryService.remove(type_id);
+            }
+            else {
+                throw new ObjectIdNotFoundException(type_id, ValueMessage.CATEGORY);
+            }
+        }
+        throw new ObjectIdValidException(type_id, ValueMessage.CATEGORY);
     }
 }
